企业数据库爬虫/Python_TagCloud-master/requirements.txt
openpyxl==2.4.9
pytagcloud==0.3.5
python_docx==0.8.6
gensim==3.2.0
xlrd==1.1.0
jieba==0.39
